added "paper" task to Makefile for full reproducibility
removed short-term hacks for quick CDDD vs XTB comparison.
seeded all operations
(even though it doesnt make a big difference,
as tested numerous times)
to make it 100% reproducible.

One thing is difficult on the conceptual level: should we
include or exclude the COSMO datapoints.
Because:
On the one hand our reviewers requested us to compare
everything against COSMO-RS as the baseline.
However, the COSMO-RS calculations fail / timeout on
many of the datapoints. Therefore, we have to heavily
restrict ourselves to only around ~30% of the dataset.
But this way, we can readily compare against the
COSMO-RS baseline without comparing apples vs oranges.

The final retrain will again consider the full dataset.