removing differential solubility for bao dataset
Due to the new bao dataset we got tremendous amount of new data.
But this creates a problem for the pairwise solubility data
that we generate for the differential solubility!

This creates just too many new datapoints, and would artificially
inflate the dataset >10^6 datapoints. Therefore, we should
remove the differential solubility for now and operate on
the absolute frame, even if this degrades performance a bit.
We can then add it in at a later stage.