implemented absolute solubilities within solvmate relative frame
the idea is that a linear solver is applied to both relative
and absolute solubilities computed with the respective
relative and absolute solubility models.
This leads two types of equations:
   a -b ... 0  = delta log S
   a  0 ... 0  = log S
that are solved by least squares concurrently. Therefore,
because there are N^2 relative solubility equations, they
correspond to a higher weight. At the same time, because
of the existence of N equations asserting the absolute scale,
we also force the relative solubilities produced to be
optimally translated into the right absolute solubilities.