Can finally do T profiles 💎 as needed for e.g. cryst experiments
The curves are all linear relationships. I concluded that the models
cannot predict phase changes. This makes sense, as the data they
are being trained on does not include phase changes...

We could try to predict phase changes by taking the insoluble parts
of e.g. the nova dataset to be phase changes. That would then allow
us to predict where precipitation is to be expected and predict
some jumps there.

This all makes a lot of sense, as the log S space is completely
smooth, and consequently modelled as a regression problem.

While, on the other hand, phase changes are necessarily discrete
events, and should consequently be modelled as binary classification.

So an idea for an additional feature would be:
color the curves by where the phase state change happens!